Solution to Check Character Type

+++ Check Character Type@@6788dd63490637cfc44e1b95/code.c
@@ -2,7 +2,7 @@
 
 int main() {
     char ch;
-    float f;
+    int f;
     
     // Take input
     printf("");
